Finished Sprint 1, making a text based GUI.
After finishing Sprint 1, I will have to Launch via filling the requirements.txt and README, and furthermore implement the changes that I listed in the review of Sprint 1, question 4. I will resort to design to the end of the task as it is a task which can be time consuming.