Finished Launch for Sprint 1. Can finally go to Sprint 2.
Made the README and requirements.txt for Sprint 1. I included ursina only because main.py was in my repository and also because in Sprint 2 you would require the ursina module anyway. Will continue to work on the main.py I have right now, as it is my most advanced file and continue to tweak it to be ready for Sprint 3.