Modularised and arranged Sprint 2 for further enhancements.
Following this, I will have to create a fully functioning prototype of the full game in mind in Sprint 2, with inbuilt UI, enemies, game modes, etc. So far I have only arranged and modularized all functions + got the sprint function to start working. Overall, this was a setup phase for my plan to finish sprint 2 with classes from the approval of the the teacher. Right now, I am questioning what to do in Sprint 3 as I am already implementing OOP fundamentals in Sprint 2.