Scrapped the idea of an enemy health bar as getting it to face the player was too tedious, did something else.
I fixed the enemy hitbox problem and how it noclipped through the ground by making a subclass of the SmoothFollow class and making the functionalities better. I also went from using a list to track bullets to raycasting to detect bullets hitting anything as it was much better for FPS games. I scrapped the idea of a healthbar and instead made the enemy cube flash red if it took damage. Overall, this is a huge fix, and will allow me to continue on to other functions, such as UI and modes.