Added a menu, encountered a new error however.
I've added a menu to my FPS game, including 4 options, starting the main game, entering free play, viewing the tutorial/instructions, and exiting the game. So far, only starting and exiting the game works, and they are working well, the menu that is. However, I encountered a new error where the bullets in the gun appear to be shaky or wobbly when I shoot them. If I can fix this error, and furthermore finish the wave based system of my game, then I would've finished a prototype of sprint 2, and can go on to add the other functionalities of the game, in sprint 3 and 4.