Fixed the gun shooting issue.
So I fixed the gun shooting issue utilising a new technique, instead of using the prebuilt entity class from ursina, I used inheritance and created a bullet class out of that entity and furthermore modified it to my liking, adding position, direction, and rotation attributes. From there, instead of the animate_position function which accounts for time instead of framerate, I used this line with the attribute i made from the bullet class: self.position += self.direction * 1000 * time.dt to control speed per frame and in what direction the bullet must move, as that line was the key conflicting factor. From now, I need to make the wave based function for the main game, and then I would've finished the prototyping for sprint 2.