Incredible changes. More to come.
Successfully made the gun from button to its own class, allowing for even more dynamic gun types. This is also another act of inheritance, greatly reducing the workload of adding guns tediously. I could do the same for enemies, however enemy logic will become too crowded. I had also tried adding a gameover menu but that was unsuccessful, resulting in many bugs after the game ended and the player was defeated. I attempted this by storing all the game entities in a list when they were created, and if the player died, the entities would all destroy themselves and the game over menu would initiate. However, this did not work, and ended up in functions with multiple arguments which was incredibly tedious to deal with. Another solution I've come up with is to simply restart the application if the player dies, however the score will not be outputted for the survival game mode. It is a quite huge factor of the use case as it is part of the post conditions, but as it seems, the goal of a game over screen is very distant.