Encapsulated many methods in my module for more clarity and structure.
Today was more or less a setup commit. I added another new method, the drop_gun method which I wasn't able to explain before. This allows you to drop the gun a few units away from the player. This will be a good feature in the future, where I will be able to equip other guns as well :) Additionally, I came with another error in the code, where if you press "Shift" and "Q" together, the game crashes. I tried doing input checks such as "if key == "q" and not held_keys["shift"] to prevent this from happening, however it still occurred. I also added some existing methods to some new classes, and eliminated the need for a preexisting FirstPersonController and Health Bar module, as I finally inherited from it and made my own classes with the methods I made. There are only 3 functions outside of classes, (Not too important), but the next primary goal is to fix the error and figure out if a game over screen is viable and to add new models, a map (Could be preexisting), and add new guns and maybe new enemies.