Fixed both errors (Kind of)
So I settled for random damage for the shotgun between 90 and 120, and then finally for the Shift + Q error, it was most likely an Ursina related error as when I tried it in other files, it also crashed. So I removed the option to swap weapons with q, and if the player clicks another weapon with a weapon equipped, they will swap it anyway. Next, I will work on more guns, and the map. If I can find a way to fix it, I will add a game over screen.