Added a working and realistic minigun
The minigun I added was simply a copy of the gun class with some minor changes. Of course, the damage is different and the fire rate as well. In the main.py, I modified the input function to check if the user is holding a minigun in particular, and holding down the left mouse button. If so, then it will initiate rapid fire. Else, if it is another type of weapon, or if they simply press the left mouse button once, then it wont work. Also some unintended features that made it a little more realistic: The gun will only work if the player is moving, meaning they will require momentum to use the gun, for proper game balancing. If the player rotates the camera too much as well, the gun will simply stop rapid firing as well, another nice feature I did not intend to add. Furthermore, a successful endeavour, and I will see to it that I add a map, and possibly a game over screen. I added a small note saying it may be possible or not to implement this feature of a game over screen.