As I was prompted firmly by the teacher to finish Sprint 3, I did so.
Overall, Sprint 2 to Sprint 3 was confusing, as I was unaware how to transition from OOP fundamentals already implemented in Sprint 2 to Sprint 3. However, Sprint 3 does contain new guns and a touch up on game logic, which I believed would be sufficient to complement Sprint 4 in the future. Sprint 4 ongoing, would be the implementation of new models and possibly shaders (If my laptop can handle it) a polished menu, a new map etc. I thought I would be able to finish the game at Sprint 3 due to my use of OOP in Sprint 2, however this was not the case, therefore removing my vision of touch ups and features rather than ingame logic in Sprint 3. I guess Sprint 3 was like a mini Sprint 4? But separating ingame logic and other features from touchups was a  good idea, and I will continue to implement these features in Sprint 4.