New Feature for Sprint 4: Health Packs
Health Packs are the new addition I decided to add to the game, as I though players may have trouble with rendering shaders. If you touch a Health Pack, it will heal 25 hp and will spawn periodically in survival mode, and you can spawn it with the r key in simulator mode. One particular issue I had with this which was different from the enemy spawning randomly function was where some of the spawning methods were encapsulated. This proved to be problematic however, because I tried to call these methods from inside the class without making an instance of the class, which I make in the method. This ended up becoming tedious, and I instead opted to handle other spawning logic and colliding logic in main.py and healing logic in module.py for simplicity. Furthermore, this was the implementation of a new feature in the game, and now begins the map and other touchups phase.