Successfully finished Sprint 1 theory today.
Finished storyboards, data flow diagram, and gantt chart accordingly. As I have done this now, I will move on to the sprint 2 theory. Soon after, I will probably not update the map extensively due to short time, and some of the gun models are not working properly, therefore I may need to revert back to the original model as I am facing problems with rotation as the bullet rotates in the opposite direction where the gun is pointed in my mini gun and pistol models, however the pistol is less evident. Therefore, I may revert to a normal gun model for the minigun and that will conclude sprint 4.