Finished the main Sprint 4 documentation. Ready to update the structure chart and the class diagram.
Also changed the Health Pack spawning to 20-50 seconds because of the short spawning times. I will soon get different textures for the map. Also due to how the hookshot allows the player to avoid any enemies, I will either remove or make the hookshot drop the player after a certain period of time.