Made protected attributes, and finished the updated UML Class Diagram.
Modified my code to include protected attributes for further encapsulation. Other attributes and methods did not require this, so I kept it to a minimum. I also updated my UML class diagram accordingly with the changes (Using # instead of + to indicate it is protected) and I also included the HealthPack class.