My final commit, I deleted unwanted assets, changed the UMLCD2 file one last time, finished off project documentation and my readme, and modified Sprint4.py and Sprint4Module.py to comply with my standards.
Made the UI more textured and changed pistol damage to be higher as it is required. Fixe the minigun and added a few more docstrings. Finished off the project documentation including the final evaluation and all charts have been updated. All extra models and assets have been deleted. Note: I will not be putting Sprint4.py and Sprint4Module.py in a folder as it removes textures from the game when it is run, therefore I am keeping it out of the folder as it is also the latest version of the game.